Добавлена авторизация портала, деление по ролям, мониторинг трендов, генерация сертификатов +qr и привязка их к аккаунту и к статье, для которой она была сгенерирована + кириллица в pdf. Также добавлена история загрузок. Подправил алгоритм формирования рекомендаций и переделал сокеты. Теперь всё должно работать лучше и правильно.
Подправил структуру БД, работал над безопасностью и правильным разграничением прав на доступ к функционалу и файлам.
Если 60 минут не было активности, сессия закроется.

+++ src/main/java/com/example/ArticleAI/models/ClassDistance.java
@@ -11,5 +11,5 @@ public class ClassDistance {
     String keyword;
     String className;
     double distance;
-    Long classActuality;
+    Double classActuality;
 }
